Add version, homepage, and schema_url to metadata schema
Additional fields have been added to the VersionedMetadataConvention class in the metadata schema. These include the version of the metadata convention, a link to its documentation (homepage), and URL of the JSON schema. Also, the 'units_v1' field is renamed to 'units' in the AllUnits class within the units schema.